type fix
git-svn-id: https://svn.code.sf.net/p/sauerbraten/code/src@280 07af1644-1164-4067-b897-63bf4dc9b748

+++ engine/ragdoll.h
@@ -351,7 +351,7 @@ void ragdolldata::move(dynent *pl, float ts)
             if(water) 
             {
                 dpos.mul(0.75f);
-                v.pos.z += 0.25f*sinf(detrnd((int)this+i, 360)*RAD + lastmillis/10000.0f*M_PI)*ts;
+                v.pos.z += 0.25f*sinf(detrnd(size_t(this)+i, 360)*RAD + lastmillis/10000.0f*M_PI)*ts;
             }
             v.pos.add(dpos);
             if(v.pos.z < 0) { v.pos.z = 0; collisions++; }
